pyproject.toml
[project]
name = "phenospex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.4",
    "open3d>=0.19.0",
    "pillow>=12.0.0",
    "plyfile>=1.1.2",
]
